% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimmaDEG.R
\name{LimmaDEG}
\alias{LimmaDEG}
\title{DEG analaysis using Limma}
\usage{
LimmaDEG(mat, anno)
}
\arguments{
\item{mat}{matrix}

\item{anno}{annotation dataframe}
}
\description{
This function performs DEG analysis using
}
